[docs] Import docsify-sidebar-collapse
This script is originated from https://github.com/iPeng6/docsify-sidebar-collapse/blob/babea8ceccdc8f585278d5fe2f88f3eef592d1f9/dist/docsify-sidebar-collapse.js

I am not the original Author. @iPeng6 is the original author for docsify-sidebar-collapse.js.

This commit is to import the script so that we can make some modifications to its behavior locally and not solely rely on it as is.