[nukerA] Implement universal module state control
We introduce a universal method for controlling the state of the antiNuker module. This is so we can unsubscribe from the playerLeave event for the onPlayerLogout function. My tests show that we can subscribe and unsubscribe to this event in a BeforeEvent however documentation and some logs shared with me by another user shows that this isn't meant to be the case for right now - subject to change in the future. In the meantime, we handle the logic differently so that unsubscribing from the playerLeave event is done so within its own event instead of the playerBreakBlockBeforeEvent.