[reachA] Clean up map when players leave
This patch will monitor players so that when they leave it can properly manage the map by cleaning up players that are no longer needed. It is also updated to be in line with standard practices to mitigate any potential reference errors for lack of system privileges by allowing to properly subscribe and unsubscribe to the playerLeave event.